Refactor create person class
