+++ options.rb
@@ -1,10 +1,12 @@
 require './student_class'
 require './teacher_class'
+require './create_person'
 require './book'
 require './rental'
 require './input'
 require './create_book'
 require './create_rental'
+
 
 # options class
 class Options
@@ -14,6 +16,7 @@ class Options
     @rentals = []
     @input = Input.new
     @create_book = CreateBook.new(@books)
+    @create_person = CreatePerson.new(@persons)
     @create_rentals = CreateRental.new(@rentals,@books,@persons)
   end
 
@@ -36,7 +39,7 @@ class Options
     when '2'
       list_people
     when '3'
-      create_person
+      @create_person.create_person
     when '4'
       @create_book.create_book
     when '5'
@@ -45,78 +48,6 @@ class Options
       list_all_rental_by_id
     end
   end
-
-  def create_person
-    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    student_or_teacher = @input.read
-    case student_or_teacher
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'Wrong Input!'
-      return
-    end
-    puts 'Person created successfully'
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = @input.read
-    print 'Name: '
-    name = @input.read
-    print 'Specialization: '
-    specialization = @input.read
-    @persons.push(Teacher.new(specialization: specialization, age: age.to_i, name: name))
-  end
-
-  def create_student
-    print 'Age: '
-    age = @input.read
-    print 'Name: '
-    name = @input.read
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = @input.read
-    unless parent_permission.upcase == 'Y' || parent_permission.upcase == 'N'
-      puts 'Wrong Input!'
-      return
-    end
-    parent_permission = parent_permission.upcase == 'Y'
-    @persons.push(Student.new(age: age.to_i, name: name, parent_permission: parent_permission))
-  end
-
-  # def create_book
-  #   print 'Title: '
-  #   title = @input.read
-  #   print 'Author: '
-  #   author = @input.read
-  #   @books.push(Book.new(title, author))
-  #   puts 'Book created successfully'
-  # end
-
-  # def create_rental
-  #   puts 'Please choose a book by number:'
-  #   @books.each_with_index do |book, index|
-  #     puts "#{index}) Title: \"#{book.title}\", Author: #{book.author}"
-  #   end
-  #   book_index = @input.read
-  #   puts 'Please choose a person from the list by number (not id)'
-  #   @persons.each_with_index do |person, index|
-  #     puts "#{index}) Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-  #   end
-
-  #   person_index = @input.read
-  #   if book_index.to_i >= @books.length || person_index.to_i >= @persons.length
-  #     puts 'Wrong Index!'
-  #     return
-  #   end
-
-  #   print 'Date: '
-  #   date = @input.read
-  #   @rentals.push(Rental.new(date, @books[book_index.to_i], @persons[person_index.to_i]))
-  #   puts 'Rental created successfully'
-  # end
 
   def list_books
     @books.each do |book|
